GND: Add a note regarding the visualization's triangle connection order
I made that visualization a long time ago, so it uses assumes an outdated (incorrect) connection order. Since the article already contains very detailed information about this issue, reworking the example is likely not needed.

Its main purpose is to illustrate how the different shapes can be combined in order to form the terrain geometry, which it does regardless of this unfortunate inaccuracy.